fix(ci): restore ha line budget
