+++ src/server/tests/alerts_and_ha.rs
@@ -2954,341 +2954,3 @@ async fn dual_active_planned_cutover_keeps_local_fenced_when_peer_response_is_am
 
     let _ = std::fs::remove_file(db_path);
 }
-
-#[tokio::test]
-async fn ha_source_endpoint_persists_origin_group_settings() {
-    let db_path = temp_db_path("ha-source-origin-group");
-    let db_str = db_path.to_string_lossy().to_string();
-    let proxy = TavilyProxy::with_endpoint(
-        vec!["tvly-ha-source-origin-group".to_string()],
-        DEFAULT_UPSTREAM,
-        &db_str,
-    )
-    .await
-    .expect("proxy created");
-    let ha = tavily_hikari::HaRuntime::new(tavily_hikari::HaConfig {
-        mode: tavily_hikari::HaMode::ActiveStandby,
-        node_id: "node-source-group".to_string(),
-        database_path: Some(db_str.clone()),
-        ..tavily_hikari::HaConfig::default()
-    });
-    let addr = spawn_ha_admin_server(proxy, ha, true).await;
-
-    let response = Client::new()
-        .put(format!("http://{addr}/api/admin/ha/source"))
-        .json(&serde_json::json!({
-            "sourceKind": "origin_group",
-            "originGroupId": "eo-group-api-test",
-            "applyToEdgeone": false
-        }))
-        .send()
-        .await
-        .expect("source settings response");
-    let status = response.status();
-    let body = response.text().await.expect("source settings body text");
-    assert!(
-        status.is_success(),
-        "source settings request should succeed, got {status}: {body}"
-    );
-    let response: Value = serde_json::from_str(&body).expect("source settings body");
-
-    assert_eq!(response["haSourceOverride"]["sourceKind"], "origin_group");
-    assert_eq!(response["haSourceOverride"]["originGroupId"], "eo-group-api-test");
-    assert_eq!(response["haSourceEffective"]["target"], "eo-group-api-test");
-    assert_eq!(response["edgeoneExpectedOrigin"], "eo-group-api-test");
-    assert_eq!(response["edgeoneExpectedSourceKind"], "origin_group");
-
-    let _ = std::fs::remove_file(db_path);
-}
-
-#[tokio::test]
-async fn ha_source_endpoint_accepts_lowercase_direct_origin_scheme() {
-    let db_path = temp_db_path("ha-source-direct-scheme");
-    let db_str = db_path.to_string_lossy().to_string();
-    let proxy = TavilyProxy::with_endpoint(
-        vec!["tvly-ha-source-direct-scheme".to_string()],
-        DEFAULT_UPSTREAM,
-        &db_str,
-    )
-    .await
-    .expect("proxy created");
-    let ha = tavily_hikari::HaRuntime::new(tavily_hikari::HaConfig {
-        mode: tavily_hikari::HaMode::ActiveStandby,
-        node_id: "node-source-direct".to_string(),
-        database_path: Some(db_str.clone()),
-        ..tavily_hikari::HaConfig::default()
-    });
-    let addr = spawn_ha_admin_server(proxy, ha, true).await;
-
-    let response = Client::new()
-        .put(format!("http://{addr}/api/admin/ha/source"))
-        .json(&serde_json::json!({
-            "sourceKind": "direct",
-            "directOriginScheme": "https",
-            "directOriginHost": "gz.ivanli.cc",
-            "directOriginPort": 1443,
-            "applyToEdgeone": false
-        }))
-        .send()
-        .await
-        .expect("source settings response");
-    let status = response.status();
-    let body = response.text().await.expect("source settings body text");
-    assert!(
-        status.is_success(),
-        "direct source settings request should succeed, got {status}: {body}"
-    );
-    let response: Value = serde_json::from_str(&body).expect("source settings body");
-
-    assert_eq!(response["haSourceOverride"]["sourceKind"], "direct");
-    assert_eq!(response["haSourceOverride"]["directOriginScheme"], "https");
-    assert_eq!(response["haSourceOverride"]["directOriginHost"], "gz.ivanli.cc");
-    assert_eq!(response["haSourceOverride"]["directOriginPort"], 1443);
-    assert_eq!(response["haSourceEffective"]["directOriginScheme"], "https");
-    assert_eq!(response["edgeoneExpectedSourceKind"], "direct");
-
-    let _ = std::fs::remove_file(db_path);
-}
-
-#[tokio::test]
-async fn ha_recovery_import_is_idempotent_and_keeps_importer_active() {
-    let db_path = temp_db_path("ha-recovery-idempotent");
-    let db_str = db_path.to_string_lossy().to_string();
-    let proxy = TavilyProxy::with_endpoint(
-        vec!["tvly-ha-recovery-idempotent".to_string()],
-        DEFAULT_UPSTREAM,
-        &db_str,
-    )
-    .await
-    .expect("proxy created");
-    let ha = tavily_hikari::HaRuntime::new(tavily_hikari::HaConfig {
-        node_id: "node-new".to_string(),
-        database_path: Some(db_str.clone()),
-        ..tavily_hikari::HaConfig::default()
-    });
-    let addr = spawn_ha_admin_server(proxy, ha, true).await;
-    let client = Client::new();
-    let payload = serde_json::json!({
-        "batchId": "old-master-batch-1",
-        "sourceNodeId": "node-old",
-        "message": "usage/log/event recovery batch imported",
-        "requestLogs": [{
-            "authTokenId": "old-token",
-            "method": "POST",
-            "path": "/api/tavily/search",
-            "statusCode": 200,
-            "tavilyStatusCode": 200,
-            "resultStatus": "success",
-            "requestKindKey": "tavily_search",
-            "requestKindLabel": "Tavily Search",
-            "requestKindDetail": "POST /api/tavily/search",
-            "businessCredits": 1,
-            "requestBody": "{\"query\":\"old-master\"}",
-            "responseBody": "{\"answer\":\"ok\"}",
-            "forwardedHeaders": "[]",
-            "droppedHeaders": "[]",
-            "visibility": "visible",
-            "createdAt": Utc::now().timestamp() - 60
-        }],
-        "authTokenLogs": [{
-            "tokenId": "old-token",
-            "method": "POST",
-            "path": "/api/tavily/search",
-            "httpStatus": 200,
-            "mcpStatus": 200,
-            "requestKindKey": "tavily_search",
-            "requestKindLabel": "Tavily Search",
-            "requestKindDetail": "POST /api/tavily/search",
-            "resultStatus": "success",
-            "countsBusinessQuota": 1,
-            "businessCredits": 1,
-            "billingState": "charged",
-            "createdAt": Utc::now().timestamp() - 60
-        }]
-    });
-
-    let rejected = client
-        .post(format!("http://{addr}/api/admin/ha/recovery/import"))
-        .json(&payload)
-        .send()
-        .await
-        .expect("rejected recovery import");
-    assert_eq!(rejected.status(), reqwest::StatusCode::BAD_REQUEST);
-    let rejected_body = rejected.text().await.expect("rejected recovery body");
-    assert!(
-        rejected_body.contains("request_logs") && rejected_body.contains("auth_token_logs"),
-        "legacy log recovery payload should be explicitly rejected: {rejected_body}"
-    );
-
-    let ledger_payload = serde_json::json!({
-        "batchId": "old-master-batch-1",
-        "sourceNodeId": "node-old",
-        "message": "ledger recovery batch imported"
-    });
-
-    let first: Value = client
-        .post(format!("http://{addr}/api/admin/ha/recovery/import"))
-        .json(&ledger_payload)
-        .send()
-        .await
-        .expect("first ledger recovery import")
-        .json()
-        .await
-        .expect("first ledger recovery response");
-    assert_eq!(first["imported"], true);
-    assert_eq!(first["eventCount"], 0);
-    assert_eq!(first["status"]["role"], "full_master");
-
-    let second: Value = client
-        .post(format!("http://{addr}/api/admin/ha/recovery/import"))
-        .json(&ledger_payload)
-        .send()
-        .await
-        .expect("second ledger recovery import")
-        .json()
-        .await
-        .expect("second ledger recovery response");
-    assert_eq!(second["imported"], false);
-
-    let pool = connect_sqlite_test_pool(&db_str).await;
-    let row: (String, i64) = sqlx::query_as(
-        "SELECT status, event_count FROM ha_recovery_batches WHERE id = 'old-master-batch-1'",
-    )
-    .fetch_one(&pool)
-    .await
-    .expect("fetch recovery batch");
-    assert_eq!(row.0, "imported");
-    assert_eq!(row.1, 0);
-    let request_log_count: i64 =
-        sqlx::query_scalar("SELECT COUNT(*) FROM request_logs WHERE auth_token_id = 'old-token'")
-            .fetch_one(&pool)
-            .await
-            .expect("fetch rejected request logs");
-    assert_eq!(request_log_count, 0);
-    let token_log_count: i64 =
-        sqlx::query_scalar("SELECT COUNT(*) FROM auth_token_logs WHERE token_id = 'old-token'")
-            .fetch_one(&pool)
-            .await
-            .expect("fetch rejected auth token logs");
-    assert_eq!(token_log_count, 0);
-    pool.close().await;
-    let _ = std::fs::remove_file(db_path);
-}
-
-#[tokio::test]
-async fn compute_signatures_tracks_recent_alert_summary_changes() {
-    let db_path = temp_db_path("summary-signatures-recent-alerts");
-    let db_str = db_path.to_string_lossy().to_string();
-    let proxy = TavilyProxy::with_endpoint(
-        vec!["tvly-signature-recent-alerts".to_string()],
-        DEFAULT_UPSTREAM,
-        &db_str,
-    )
-    .await
-    .expect("proxy created");
-
-    let user = proxy
-        .upsert_oauth_account(&OAuthAccountProfile {
-            provider: "linuxdo".to_string(),
-            provider_user_id: "linuxdo-signature-alert-user".to_string(),
-            username: Some("sig_alert".to_string()),
-            name: Some("Sig Alert".to_string()),
-            avatar_template: None,
-            active: true,
-            trust_level: Some(1),
-            raw_payload_json: None,
-        })
-        .await
-        .expect("upsert oauth user");
-    let token = proxy
-        .ensure_user_token_binding(&user.user_id, Some("signature-alert-bound"))
-        .await
-        .expect("ensure token binding");
-
-    let state = Arc::new(AppState {
-        proxy: proxy.clone(),
-        static_dir: None,
-        forward_auth: ForwardAuthConfig::new(None, None, None, None),
-        forward_auth_enabled: false,
-        builtin_admin: BuiltinAdminAuth::new(false, None, None),
-        linuxdo_oauth: LinuxDoOAuthOptions::disabled(),
-        linuxdo_credit: LinuxDoCreditOptions::disabled(),
-        ha: tavily_hikari::HaRuntime::new(tavily_hikari::HaConfig::default()),
-        dev_open_admin: false,
-        usage_base: "http://127.0.0.1:58088".to_string(),
-        api_key_ip_geo_origin: "https://api.country.is".to_string(),
-        dashboard_overview_cache: new_dashboard_overview_cache(),
-    });
-
-    let (before_sig, _) = compute_signatures(&state)
-        .await
-        .expect("compute signatures before alerts");
-    let before_sig = before_sig.expect("summary signature before alerts");
-    assert_eq!(before_sig.freshness.recent_alerts_total_events, 0);
-    assert_eq!(before_sig.freshness.recent_alerts_grouped_count, 0);
-
-    let now = Utc::now().timestamp();
-    let pool = connect_sqlite_test_pool(&db_str).await;
-    sqlx::query(
-            r#"
-            INSERT INTO auth_token_logs (
-                token_id,
-                method,
-                path,
-                query,
-                http_status,
-                mcp_status,
-                request_kind_key,
-                request_kind_label,
-                request_kind_detail,
-                result_status,
-                error_message,
-                key_effect_code,
-                binding_effect_code,
-                selection_effect_code,
-                counts_business_quota,
-                created_at
-            ) VALUES (?, 'POST', '/mcp', NULL, 429, -1, 'mcp_search', 'MCP Search', 'POST /mcp', 'quota_exhausted', 'hourly any-request limit exceeded', 'none', 'none', 'none', 0, ?)
-            "#,
-        )
-        .bind(&token.id)
-        .bind(now)
-        .execute(&pool)
-        .await
-        .expect("insert recent alert auth token log");
-
-    let (after_sig, _) = compute_signatures(&state)
-        .await
-        .expect("compute signatures after alerts");
-    let after_sig = after_sig.expect("summary signature after alerts");
-    assert_eq!(after_sig.freshness.recent_alerts_total_events, 1);
-    assert_eq!(after_sig.freshness.recent_alerts_grouped_count, 1);
-    assert_eq!(
-        after_sig.freshness.recent_alerts_counts,
-        vec![
-            (
-                tavily_hikari::ALERT_TYPE_UPSTREAM_RATE_LIMITED_429.to_string(),
-                0
-            ),
-            (
-                tavily_hikari::ALERT_TYPE_UPSTREAM_USAGE_LIMIT_432.to_string(),
-                0
-            ),
-            (
-                tavily_hikari::ALERT_TYPE_UPSTREAM_KEY_BLOCKED.to_string(),
-                0
-            ),
-            (
-                tavily_hikari::ALERT_TYPE_USER_REQUEST_RATE_LIMITED.to_string(),
-                1
-            ),
-            (
-                tavily_hikari::ALERT_TYPE_USER_QUOTA_EXHAUSTED.to_string(),
-                0
-            ),
-        ]
-    );
-    assert_ne!(before_sig, after_sig);
-
-    let _ = std::fs::remove_file(db_path);
-}
