+++ src/store/mod.rs
@@ -2850,6 +2850,7 @@ pub(crate) struct KeyStore {
 }
 
 include!("key_store_bootstrap.rs");
+include!("key_store_bootstrap_legacy.rs");
 include!("key_store_quota_schema_semantic_migration.rs");
 include!("key_store_observability_sidecar.rs");
 include!("key_store_public_metrics_freshness.rs");
@@ -2879,6 +2880,7 @@ include!("key_store_token_success_metrics.rs");
 include!("key_store_jobs.rs");
 include!("key_store_account_limit_snapshots.rs");
 include!("key_store_account_usage_rollups.rs");
+include!("key_store_ha_defs.rs");
 include!("key_store_ha.rs");
 #[cfg(test)]
 include!("key_store_request_logs_and_dashboard_test_support.rs");
